+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-09-03T07:45:51.499Z",
+    "2026-09-03T10:05:12.570Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {